Incorrect order of arguments (#504)
It seems that energy_balance() expects that the order in the var_name argument is (varin, varout, varenergy). Here it was called with (varout, varin, varenergy). If the efficiency data is not all 100% this results in a wrong constraint. 

For example considering efficiency (in = 0.9, out = 0.8) the current constraint will look like:  
E__GenericBattery_{313_HEAD_STORAGE, 3} - E__GenericBattery_{313_HEAD_STORAGE, 2} - 0.9 Pout__GenericBattery_{313_HEAD_STORAGE, 3} + 1.25 Pin__GenericBattery_{313_HEAD_STORAGE, 3} = 0.0
The 1.25 times addition of the charging power to the energy is faulty.